--- a/src/net/http/fs.go
+++ b/src/net/http/fs.go
@@ -333,7 +333,7 @@ func serveContent(w ResponseWriter, r *Request, name string, modtime time.Time, 
 
 	w.WriteHeader(code)
 
-	if r.Method != "HEAD" {
+	if r.Method != MethodHead {
 		if sendSize == size {
 			// use Copy in the non-range case to make use of WriterTo if available
 			io.Copy(w, sendContent)