This small PR allows optimizations made in `io.Copy` (like the use of `io.WriterTo`) to be used in one possible path of `http.ServeContent` (in case of a non-`Range` request).
This, in turn, allows us to skip the buffer allocation in `io.Copy`, which has performance benefits:

```
name            old time/op    new time/op    delta
ServeContent-8    39.0µs ±10%    22.2µs ± 6%  -43.17%  (p=0.000 n=9+10)

name            old alloc/op   new alloc/op   delta
ServeContent-8    32.9kB ± 0%     0.1kB ± 0%  -99.80%  (p=0.000 n=8+7)

name            old allocs/op  new allocs/op  delta
ServeContent-8      6.00 ± 0%      4.00 ± 0%  -33.33%  (p=0.000 n=10+10)
```

(results are only representative for non-`Range` requests; `Range` requests see no changes in performance)

<details>
<summary>Benchmark code</summary>

```go
func BenchmarkServeContent(b *testing.B) {
	type readSeekWriterTo interface {
		io.ReadSeeker
		io.WriterTo
	}

	content, err := os.ReadFile("../../testdata/Isaac.Newton-Opticks.txt")
	if err != nil {
		b.Fatalf("could not open source file: %v", err)
	}

	in := struct{ readSeekWriterTo }{bytes.NewReader(content)}

	req := httptest.NewRequest(MethodGet, "/", NoBody)
	// req.Header.Set("Range", "bytes=0-100000")
	rw := httptest.NewRecorder()
	rw.Header().Set("Content-Type", "text/plain")

	b.ReportAllocs()
	b.ResetTimer()
	for i := 0; i <= b.N; i++ {
		in.Seek(0, io.SeekStart)
		ServeContent(rw, req, "foo.txt", time.Time{}, in)
		rw.Body.Reset()
	}
}
```

</details>