@costela 2022-10-28
Benchmark code: https://gist.github.com/costela/44079c1c742007d355cb3e382d0ac0ef